python_scripts/requirements.txt
anyio==3.6.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
deprecation==2.1.0
gotrue==0.5.0
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
idna==3.3
packaging==21.3
postgrest-py==0.10.2
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
realtime==0.0.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
storage3==0.3.2
supabase==0.5.6
typing_extensions==4.2.0
websockets==9.1
